+++ src/main/java/ru/moysklad/remap_1_2/entities/permissions/value/BaseEntityPermissionValue.java
@@ -1,0 +1,17 @@
+package ru.moysklad.remap_1_2.entities.permissions.value;
+
+import lombok.EqualsAndHashCode;
+import lombok.Getter;
+import lombok.NoArgsConstructor;
+import lombok.Setter;
+
+@Getter
+@Setter
+@NoArgsConstructor
+@EqualsAndHashCode
+public class BaseEntityPermissionValue {
+    private EntityPermissionValue view;
+    private EntityPermissionValue create;
+    private EntityPermissionValue update;
+    private EntityPermissionValue delete;
+}