+++ src/main/java/ru/moysklad/remap_1_2/entities/permissions/value/BaseScriptPermissionValue.java
@@ -1,0 +1,9 @@
+package ru.moysklad.remap_1_2.entities.permissions.value;
+
+public enum BaseScriptPermissionValue {
+    NO,
+    AUTHOR,
+    ASSIGNEE,
+    AUTHOR_OR_ASSIGNEE,
+    ALL
+}