+++ src/main/java/ru/moysklad/remap_1_2/entities/permissions/value/EntityPermissionValue.java
@@ -1,0 +1,10 @@
+package ru.moysklad.remap_1_2.entities.permissions.value;
+
+public enum EntityPermissionValue {
+    NO,
+    OWN,
+    OWN_SHARED,
+    OWN_GROUP,
+    OWN_GROUP_SHARED,
+    ALL
+}